+++ packages/docs/demos/new_components/ChipDemo.vue
@@ -17,6 +17,22 @@ const handleDelete = () => {
     <Row :gap="8">
       <Chip label="Basic Chip" />
       <Chip label="With Delete" deletable @delete="handleDelete" />
+      <Chip label="With Deletate & Avatar" deletable>
+        <template #avatar>
+          <div
+            style="
+              background: #2196f3;
+              width: 100%;
+              height: 100%;
+              display: flex;
+              align-items: center;
+              justify-content: center;
+            "
+          >
+            <span style="color: white; font-size: 10px">A</span>
+          </div>
+        </template>
+      </Chip>
       <Chip label="With Avatar">
         <template #avatar>
           <div