+++ packages/fluekit/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+## [2.2.4] - 2026-01-22
+
+### Features
+
+- **Chip**:
+  - Add `size` prop support (`small`, `medium`, `large`).
+  - Add `width` and `height` props for fixed dimensions.
+  - Automatically center content when fixed dimensions are used.
+
 ## [2.2.3] - 2026-01-22
 
 ### Features